+++ tests/utils/card.py
@@ -10,11 +10,11 @@ example_card = IssuingCard(
 )
 
 
-def generateExampleCardsJson(holder, n=1, type=None, bin_id=None):
+def generateExampleCardsJson(holder, n=1, type=None, product_id=None):
     cards = []
     for _ in range(n):
-        if bin_id:
-            example_card.bin_id = bin_id
+        if product_id:
+            example_card.product_id = product_id
         example_card.holder_name = holder.name
         example_card.holder_tax_id = holder.tax_id
         example_card.holder_external_id = holder.external_id