+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 ## [Unreleased]
 ### Added
 - product_id attribute to IssuingPurchase resource
+- type attribute to IssuingDesign resource
 
 ## [0.5.0] - 2022-11-23
 ### Changed